Test updated text provided by getCatFactPromise
After the Promise from getCatFactPromise() resolves, the result is
updated in state and the resulting render changes.

Since this is not the initial rendering, we need to wrap

getByText('fact from promise')

in a function in await waitFor() to repeat the text until the update
occurs.

Note: There is a limit to how long the code will wait (e.g. if you add a
setTimeout() to the Promise before it resolves). In my experience, 1
second worked but 2 or higher failed.

See #13